fix(frontend): Slice 7 review corrections - test-only probe, definitive import rejection, JS-edit lifecycle, JS-less symbol gate
Addresses the four Slice 7 review gaps.

1. The authority probe is now genuinely test-only. Moved to
   stdlib/cli/js/hull/source/tests/frontend_probe.js, which the production cli-js
   registry generator prunes (it excludes any tests directory), so no shipped
   binary embeds it. A separate test-only registry (STDLIB_JS_CLI_TEST_REGISTRY_O,
   same array symbol) includes it and is linked only by test_js_generation. Added a
   shipped-registry assertion (grep 'frontend_probe' on $HULL, both binaries) that
   fails if any probe module is embedded.

2. The fake-import claim is now definitive. Replaced the weak in-JS dynamic-import
   outcome with a C test-only hl_js_gen_probe_import that loads an arbitrary module
   through the real tooling loader. The test asserts a fake application-module name
   fails with the loader's specific "tooling entry module not found" (a
   resolved-but-wrong module yields a DIFFERENT message, so the test fails if the
   name ever resolves), and that the real probe module still loads through the same
   path.

3. The mixed lifecycle now edits JavaScript. Before stopping dev, the mixed leg
   records the generation, edits client.js (adds an annotated handler2), waits for a
   strictly higher generation, asserts the changed JS declaration + its new @route
   annotation appear, and uses that final live document for the standalone semantic
   comparison. C5 is re-proven with JS present.

4. The JS-less symbol assertion is now in the gate. e2e-project-discovery-lua runs a
   fail-closed nm check (zero hl_js_gen_ / QuickJS symbols) before the lifecycle
   script.

Clean make test is green; the transient local test_js failure was config-flip stale
artifacts (isolated per-job in CI), not a defect.
